feat: add the ESP-IDF layer, Matter integration and the documentation set
- ac_hal: SPS30 over SHDLC/UART, SGP40 low-power VOC, SCD41 power-cycled
  single shot, SSD1681 e-paper, MAX17048, button, NVS
- Matter: air quality, temperature, humidity and power source endpoints, all
  standard device types, SIT ICD at a 15 s poll
- vendored Sensirion Gas Index Algorithm (BSD-3)
- README plus thirteen documents, including an honest testing status and a
  risks document that says what this device cannot do

+++ firmware/components/ac_core/src/ac_display.c
@@ -293,7 +293,8 @@ static void screen_system(ac_canvas_t *c, const ac_engine_t *e,
     kv_row(c, 138, "THREAD", ui->thread_attached ? "UP" : "DOWN");
     kv_row(c, 158, "MATTER", ui->matter_commissioned ? "PAIRED" : "NOT PAIRED");
     uint32_t up = ac_engine_uptime_s(e, now);
-    snprintf(v, sizeof(v), "%ud %02uh", up / 86400u, (up % 86400u) / 3600u);
+    snprintf(v, sizeof(v), "%ud %02uh", (unsigned)(up / 86400u),
+             (unsigned)((up % 86400u) / 3600u));
     /* shares the hero label's line, right aligned, so it cannot collide */
     ac_canvas_text_right(c, &ac_font_small, AC_DISP_W - GRID, 44, v, true);
     status_strip(c, e, ui);
